worker/requirements.txt
requests==2.32.3
SQLAlchemy==2.0.36
pymysql==1.1.1
python-dateutil==2.9.0.post0
